+++ src/components/layout/header/header.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import clsx from 'clsx';
 import { observer } from 'mobx-react-lite';
-import { standalone_routes } from '@/components/shared';
+import { generateOAuthURL, standalone_routes } from '@/components/shared';
 import Button from '@/components/shared_ui/button';
 import useActiveAccount from '@/hooks/api/account/useActiveAccount';
 import { useApiBase } from '@/hooks/useApiBase';
@@ -18,7 +18,6 @@ import CustomNotifications from './custom-notifications';
 import MenuItems from './menu-items';
 import MobileMenu from './mobile-menu';
 import PlatformSwitcher from './platform-switcher';
-import { generateOAuthURL } from './utils';
 import './header.scss';
 
 const { getOauthURL } = URLUtils;
@@ -70,8 +69,7 @@ const AppHeader = observer(() => {
                     <Button
                         tertiary
                         onClick={() => {
-                            const oauth_url = generateOAuthURL();
-                            window.location.replace(oauth_url);
+                            window.location.replace(generateOAuthURL());
                         }}
                     >
                         <Localize i18n_default_text='Log in' />
