+++ src/components/layout/header/utils/index.js
@@ -1,19 +1,4 @@
 import { api_base } from '@/external/bot-skeleton';
-import { LocalStorageConstants,LocalStorageUtils, URLUtils } from '@deriv-com/utils';
-
-export const generateOAuthURL = () => {
-    const { getOauthURL } = URLUtils;
-    const oauth_url = getOauthURL();
-    const original_url = new URL(oauth_url);
-    const configured_server_url =
-        LocalStorageUtils.getValue(LocalStorageConstants.configServerURL) || original_url.hostname;
-
-    const validServerURLs = ['green.derivws.com', 'red.derivws.com', 'blue.derivws.com'];
-    if (!validServerURLs.includes(configured_server_url)) {
-        original_url.hostname = configured_server_url;
-    }
-    return original_url.toString() || oauth_url;
-};
 
 export const LOW_RISK_COUNTRIES = () => ['za', 'ec', 'bw'];
 export const EU_COUNTRIES = () => [
