+++ src/components/shared/utils/config/config.ts
@@ -1,3 +1,4 @@
+import { LocalStorageConstants, LocalStorageUtils, URLUtils } from '@deriv-com/utils';
 import { isStaging } from '../url/helpers';
 
 export const APP_IDS = {
@@ -135,4 +136,18 @@ export const getDebugServiceWorker = () => {
     if (debug_service_worker_flag) return !!parseInt(debug_service_worker_flag);
 
     return false;
+};
+
+export const generateOAuthURL = () => {
+    const { getOauthURL } = URLUtils;
+    const oauth_url = getOauthURL();
+    const original_url = new URL(oauth_url);
+    const configured_server_url = (LocalStorageUtils.getValue(LocalStorageConstants.configServerURL) ||
+        original_url.hostname) as string;
+
+    const validServerURLs = ['green.derivws.com', 'red.derivws.com', 'blue.derivws.com'];
+    if (!validServerURLs.includes(configured_server_url)) {
+        original_url.hostname = configured_server_url;
+    }
+    return original_url.toString() || oauth_url;
 };
